Try harder to unmount devices in test cleanup
Unmount may fail with "Device or resource busy" so just try it
again if needed.
Also move the mount and unmount functions to test utils.
